feat(tuya-panel-szos-sdk): add test case
